feat: Change virtualenvwrapper source order in bash
Previously, I was sourcing the virtualenvwrapper bash script as part of
the base runcom settings. This was causing errors because sometimes we
change the PATH—and thus the python version/binary—in the OS-specific
setting files.

This change adds the concept of a "post-os" settings file that runs
between the OS and local settings. It additionally forces OSX to use the
Homebrew python.